Fix superb prop task; add support for smearing colorvecs.

+++ lib/meas/inline/hadron/inline_prop_and_matelem_distillation_superb_w.h
@@ -46,7 +46,7 @@ namespace Chroma
 	  bool          zero_colorvecs;
 	  bool          fuse_timeloop;  
           int           max_rhs;        /*! maximum number of linear systems solved simultaneously */
-	  multi1d<int>  phase;          /*!< Phase to apply to colorvecs */
+	  multi1d<float> phase;         /*!< Phase to apply to colorvecs */
 	};
 
 	ChromaProp_t    prop;
